Added entity class
